+++ docker/stop.sh
@@ -3,7 +3,7 @@ if [ "$#" -ne 1 ]; then
 	exit 1
 fi
 
-echo "noVNC-$1 running on port $((8081 + $1)) stopped."
+echo "noVNC-$1 running on port $((3001 + $1)) stopped."
 
 docker container kill noVNC-$1